feat: Add ability to include only LIBRARY or UI stuff
