+++ src/Uno.UI.RuntimeTests.Engine.Library/Library/RequiresFullWindowAttribute.cs
@@ -4,7 +4,7 @@ using System;
 
 namespace Uno.UI.RuntimeTests;
 
-#if !UNO_RUNTIMETESTS_DISABLE_REQUIRESFULLWINDOWATTRIBUTE
+#if !UNO_RUNTIMETESTS_DISABLE_LIBRARY && !UNO_RUNTIMETESTS_DISABLE_REQUIRESFULLWINDOWATTRIBUTE
 /// <summary>
 /// Marks a test which sets its test UI to be full-screen (Window.Content).
 /// </summary>
