+++ src/Uno.UI.RuntimeTests.Engine.Library/Library/RunsOnUIThreadAttribute.cs
@@ -4,7 +4,7 @@ using System;
 
 namespace Uno.UI.RuntimeTests;
 
-#if !UNO_RUNTIMETESTS_DISABLE_RUNSONUITHREADATTRIBUTE
+#if !UNO_RUNTIMETESTS_DISABLE_LIBRARY && !UNO_RUNTIMETESTS_DISABLE_RUNSONUITHREADATTRIBUTE
 /// <summary>
 /// Indicates that the test should be run on the UI thread.
 /// </summary>
