--- src/Uno.UI.RuntimeTests.Engine.Library/TestRun.cs
@@ -1,15 +1,0 @@
-﻿namespace Uno.UI.RuntimeTests
-{
-	public sealed partial class UnitTestsControl
-	{
-		private class TestRun
-		{
-			public int Run { get; set; }
-			public int Ignored { get; set; }
-			public int Succeeded { get; set; }
-			public int Failed { get; set; }
-
-			public int CurrentRepeatCount { get; set; }
-		}
-	}
-}
