+++ src/Uno.UI.RuntimeTests.Engine.Library/UI/TestCaseResult.cs
@@ -5,6 +5,8 @@ using System.Linq;
 
 namespace Uno.UI.RuntimeTests;
 
+#if !UNO_RUNTIMETESTS_DISABLE_UI
+
 internal record TestCaseResult
 {
 	public TestResult TestResult { get; init; }
@@ -12,3 +14,6 @@ internal record TestCaseResult
 	public TimeSpan Duration { get; init; }
 	public string? Message { get; init; }
 }
+
+
+#endif
