+++ src/Uno.UI.RuntimeTests.Engine.Library/UI/TestResult.cs
@@ -3,6 +3,8 @@ using System.Linq;
 
 namespace Uno.UI.RuntimeTests;
 
+#if !UNO_RUNTIMETESTS_DISABLE_UI
+
 internal enum TestResult
 {
 	Passed,
@@ -10,3 +12,5 @@ internal enum TestResult
 	Error,
 	Skipped,
 }
+
+#endif
