+++ src/Uno.UI.RuntimeTests.Engine.Library/UI/TestRun.cs
@@ -1,0 +1,18 @@
+﻿namespace Uno.UI.RuntimeTests;
+
+#if !UNO_RUNTIMETESTS_DISABLE_UI
+
+public sealed partial class UnitTestsControl
+{
+	private class TestRun
+	{
+		public int Run { get; set; }
+		public int Ignored { get; set; }
+		public int Succeeded { get; set; }
+		public int Failed { get; set; }
+
+		public int CurrentRepeatCount { get; set; }
+	}
+}
+
+#endif
